+++ builder/amazon/common/block_device_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/aws/awsutil"
 	"github.com/aws/aws-sdk-go/service/ec2"
 )
 
@@ -17,22 +16,20 @@ func TestBlockDevice(t *testing.T) {
 		{
 			Config: &BlockDevice{
 				DeviceName:          "/dev/sdb",
-				VirtualName:         "ephemeral0",
 				SnapshotId:          "snap-1234",
 				VolumeType:          "standard",
 				VolumeSize:          8,
 				DeleteOnTermination: true,
-				IOPS:                1000,
 			},
 
 			Result: &ec2.BlockDeviceMapping{
 				DeviceName:  aws.String("/dev/sdb"),
-				VirtualName: aws.String("ephemeral0"),
-				EBS: &ec2.EBSBlockDevice{
-					SnapshotID:          aws.String("snap-1234"),
+				VirtualName: aws.String(""),
+				Ebs: &ec2.EbsBlockDevice{
+					SnapshotId:          aws.String("snap-1234"),
 					VolumeType:          aws.String("standard"),
-					VolumeSize:          aws.Long(8),
-					DeleteOnTermination: aws.Boolean(true),
+					VolumeSize:          aws.Int64(8),
+					DeleteOnTermination: aws.Bool(true),
 				},
 			},
 		},
@@ -45,17 +42,16 @@ func TestBlockDevice(t *testing.T) {
 			Result: &ec2.BlockDeviceMapping{
 				DeviceName:  aws.String("/dev/sdb"),
 				VirtualName: aws.String(""),
-				EBS: &ec2.EBSBlockDevice{
+				Ebs: &ec2.EbsBlockDevice{
 					VolumeType:          aws.String(""),
-					VolumeSize:          aws.Long(8),
-					DeleteOnTermination: aws.Boolean(false),
+					VolumeSize:          aws.Int64(8),
+					DeleteOnTermination: aws.Bool(false),
 				},
 			},
 		},
 		{
 			Config: &BlockDevice{
 				DeviceName:          "/dev/sdb",
-				VirtualName:         "ephemeral0",
 				VolumeType:          "io1",
 				VolumeSize:          8,
 				DeleteOnTermination: true,
@@ -64,13 +60,24 @@ func TestBlockDevice(t *testing.T) {
 
 			Result: &ec2.BlockDeviceMapping{
 				DeviceName:  aws.String("/dev/sdb"),
-				VirtualName: aws.String("ephemeral0"),
-				EBS: &ec2.EBSBlockDevice{
+				VirtualName: aws.String(""),
+				Ebs: &ec2.EbsBlockDevice{
 					VolumeType:          aws.String("io1"),
-					VolumeSize:          aws.Long(8),
-					DeleteOnTermination: aws.Boolean(true),
-					IOPS:                aws.Long(1000),
+					VolumeSize:          aws.Int64(8),
+					DeleteOnTermination: aws.Bool(true),
+					Iops:                aws.Int64(1000),
 				},
+			},
+		},
+		{
+			Config: &BlockDevice{
+				DeviceName:  "/dev/sdb",
+				VirtualName: "ephemeral0",
+			},
+
+			Result: &ec2.BlockDeviceMapping{
+				DeviceName:  aws.String("/dev/sdb"),
+				VirtualName: aws.String("ephemeral0"),
 			},
 		},
 	}
@@ -84,11 +91,14 @@ func TestBlockDevice(t *testing.T) {
 		expected := []*ec2.BlockDeviceMapping{tc.Result}
 		got := blockDevices.BuildAMIDevices()
 		if !reflect.DeepEqual(expected, got) {
-			t.Fatalf("Bad block device, \nexpected: %s\n\ngot: %s", awsutil.StringValue(expected), awsutil.StringValue(got))
+			t.Fatalf("Bad block device, \nexpected: %s\n\ngot: %s",
+				expected, got)
 		}
 
 		if !reflect.DeepEqual(expected, blockDevices.BuildLaunchDevices()) {
-			t.Fatalf("Bad block device, \nexpected: %s\n\ngot: %s", awsutil.StringValue(expected), awsutil.StringValue(blockDevices.BuildLaunchDevices()))
+			t.Fatalf("Bad block device, \nexpected: %s\n\ngot: %s",
+				expected,
+				blockDevices.BuildLaunchDevices())
 		}
 	}
 }