+++ builder/amazon/common/step_create_tags.go
@@ -34,9 +34,14 @@ func (s *StepCreateTags) Run(state multistep.StateBag) multistep.StepAction {
 			// Declare list of resources to tag
 			resourceIds := []*string{&ami}
 
+			regionconn := ec2.New(&aws.Config{
+				Credentials: ec2conn.Config.Credentials,
+				Region:      aws.String(region),
+			})
+
 			// Retrieve image list for given AMI
-			imageResp, err := ec2conn.DescribeImages(&ec2.DescribeImagesInput{
-				ImageIDs: resourceIds,
+			imageResp, err := regionconn.DescribeImages(&ec2.DescribeImagesInput{
+				ImageIds: resourceIds,
 			})
 
 			if err != nil {
@@ -57,16 +62,11 @@ func (s *StepCreateTags) Run(state multistep.StateBag) multistep.StepAction {
 
 			// Add only those with a Snapshot ID, i.e. not Ephemeral
 			for _, device := range image.BlockDeviceMappings {
-				if device.EBS != nil && device.EBS.SnapshotID != nil {
-					ui.Say(fmt.Sprintf("Tagging snapshot: %s", *device.EBS.SnapshotID))
-					resourceIds = append(resourceIds, device.EBS.SnapshotID)
+				if device.Ebs != nil && device.Ebs.SnapshotId != nil {
+					ui.Say(fmt.Sprintf("Tagging snapshot: %s", *device.Ebs.SnapshotId))
+					resourceIds = append(resourceIds, device.Ebs.SnapshotId)
 				}
 			}
-
-			regionconn := ec2.New(&aws.Config{
-				Credentials: ec2conn.Config.Credentials,
-				Region:      region,
-			})
 
 			_, err = regionconn.CreateTags(&ec2.CreateTagsInput{
 				Resources: resourceIds,