+++ builder/amazon/common/step_modify_ami_attributes.go
@@ -44,21 +44,35 @@ func (s *StepModifyAMIAttributes) Run(state multistep.StateBag) multistep.StepAc
 
 	if len(s.Groups) > 0 {
 		groups := make([]*string, len(s.Groups))
+		adds := make([]*ec2.LaunchPermission, len(s.Groups))
+		addGroups := &ec2.ModifyImageAttributeInput{
+			LaunchPermission: &ec2.LaunchPermissionModifications{},
+		}
+
 		for i, g := range s.Groups {
-			groups[i] = &g
+			groups[i] = aws.String(g)
+			adds[i] = &ec2.LaunchPermission{
+				Group: aws.String(g),
+			}
 		}
-		options["groups"] = &ec2.ModifyImageAttributeInput{
-			UserGroups: groups,
-		}
+		addGroups.UserGroups = groups
+		addGroups.LaunchPermission.Add = adds
+
+		options["groups"] = addGroups
 	}
 
 	if len(s.Users) > 0 {
 		users := make([]*string, len(s.Users))
+		adds := make([]*ec2.LaunchPermission, len(s.Users))
 		for i, u := range s.Users {
-			users[i] = &u
+			users[i] = aws.String(u)
+			adds[i] = &ec2.LaunchPermission{UserId: aws.String(u)}
 		}
 		options["users"] = &ec2.ModifyImageAttributeInput{
-			UserIDs: users,
+			UserIds: users,
+			LaunchPermission: &ec2.LaunchPermissionModifications{
+				Add: adds,
+			},
 		}
 	}
 
@@ -76,11 +90,11 @@ func (s *StepModifyAMIAttributes) Run(state multistep.StateBag) multistep.StepAc
 		ui.Say(fmt.Sprintf("Modifying attributes on AMI (%s)...", ami))
 		regionconn := ec2.New(&aws.Config{
 			Credentials: ec2conn.Config.Credentials,
-			Region:      region,
+			Region:      aws.String(region),
 		})
 		for name, input := range options {
 			ui.Message(fmt.Sprintf("Modifying: %s", name))
-			input.ImageID = &ami
+			input.ImageId = &ami
 			_, err := regionconn.ModifyImageAttribute(input)
 			if err != nil {
 				err := fmt.Errorf("Error modify AMI attributes: %s", err)